+++ kleidicv_thread/include/kleidicv_thread/kleidicv_thread.h
@@ -339,6 +339,46 @@ kleidicv_error_t kleidicv_thread_sobel_3x3_horizontal_s16_u8(
 
 /// Internal - not part of the public API and its direct use is not supported.
 ///
+/// Multithreaded implementation of kleidicv_median_blur_u8 - see
+/// the documentation of that function for more details.
+kleidicv_error_t kleidicv_thread_median_blur_u8(
+    const uint8_t *src, size_t src_stride, uint8_t *dst, size_t dst_stride,
+    size_t width, size_t height, size_t channels, size_t kernel_width,
+    size_t kernel_height, kleidicv_border_type_t border_type,
+    kleidicv_thread_multithreading);
+
+/// Internal - not part of the public API and its direct use is not supported.
+///
+/// Multithreaded implementation of kleidicv_median_blur_s16 - see
+/// the documentation of that function for more details.
+kleidicv_error_t kleidicv_thread_median_blur_s16(
+    const int16_t *src, size_t src_stride, int16_t *dst, size_t dst_stride,
+    size_t width, size_t height, size_t channels, size_t kernel_width,
+    size_t kernel_height, kleidicv_border_type_t border_type,
+    kleidicv_thread_multithreading mt);
+
+/// Internal - not part of the public API and its direct use is not supported.
+///
+/// Multithreaded implementation of kleidicv_median_blur_u16 - see
+/// the documentation of that function for more details.
+kleidicv_error_t kleidicv_thread_median_blur_u16(
+    const uint16_t *src, size_t src_stride, uint16_t *dst, size_t dst_stride,
+    size_t width, size_t height, size_t channels, size_t kernel_width,
+    size_t kernel_height, kleidicv_border_type_t border_type,
+    kleidicv_thread_multithreading mt);
+
+/// Internal - not part of the public API and its direct use is not supported.
+///
+/// Multithreaded implementation of kleidicv_median_blur_f32 - see
+/// the documentation of that function for more details.
+kleidicv_error_t kleidicv_thread_median_blur_f32(
+    const float *src, size_t src_stride, float *dst, size_t dst_stride,
+    size_t width, size_t height, size_t channels, size_t kernel_width,
+    size_t kernel_height, kleidicv_border_type_t border_type,
+    kleidicv_thread_multithreading mt);
+
+/// Internal - not part of the public API and its direct use is not supported.
+///
 /// Multithreaded implementation of kleidicv_sobel_3x3_vertical_s16_u8 - see the
 /// documentation of that function for more details.
 kleidicv_error_t kleidicv_thread_sobel_3x3_vertical_s16_u8(