+++ scripts/ci-opencv.sh
@@ -114,6 +114,7 @@ IMGPROC_TEST_PATTERNS=(
   '*Imgproc_Erode*'
   '*Imgproc_PyramidDown*'
   '*Imgproc_Remap*'
+  '*Imgproc_MedianBlur*'
   '*Imgproc_Warp*'
 )
 IMGPROC_TEST_PATTERNS_STR="$(join_strings_with_colon "${IMGPROC_TEST_PATTERNS[*]}")"