+++ CHANGELOG.md
@@ -15,6 +15,9 @@ This changelog aims to follow the guiding principles of
 
 ## 0.5.0 - not yet released
 
+### Added
+- Median Blur for 5x5 kernels.
+
 ## 0.4.0 - 2025-03-25
 
 ### Added